refactor(frontend): split useKnowledgeBase into 6 domain composables + drop dead try/catch boilerplate (#5122 #5123)
